ToDoAPI/backend/accounts/urls.py
from rest_framework_simplejwt.views import TokenObtainPairView
from django.urls import path
from .views import RegisterView

urlpatterns = [
    path('register/', RegisterView.as_view(), name='register'),
    path('login/', TokenObtainPairView.as_view(), name='login')

]
