ToDoAPI/backend/requirements.txt
asgiref==3.8.1
autopep8==2.3.2
certifi==2025.1.31
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
Faker==37.0.2
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pillow==11.1.0
pluggy==1.5.0
pycodestyle==2.12.1
PyJWT==1.7.1
pytest==8.3.5
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
